refactor: 번다운 차트 위젯을 별도의 코드로 분리 (CM-172)
